refactor: 도메인 타입을 src/types/에서 src/api/*.types로 이동
API가 데이터 계약의 소유자가 되도록 menu/notice/vote 타입을
src/api/ 하위로 이동하고 전체 import 경로를 갱신한다.
