+++ src/api/getAnnouncements.ts
@@ -1,6 +1,6 @@
 import {ANNOUNCEMENTS} from '@/data/announcements';
 import dayjs from '@/lib/dayjs';
-import type {Notice} from '@/types/notice';
+import type {Notice} from '@/api/notice.types';
 
 export const getAnnouncements = (): Notice[] => {
   const now = dayjs();
