+++ src/api/notice.types.ts
