+++ src/api/vote.types.ts
