--- src/types/meal.ts
@@ -1,1 +1,0 @@
-export type MealType = 'BREAKFAST' | 'LUNCH' | 'DINNER';
